Revise the flooding protocol, also merging the relative nodetypes etc in the flood.rs file and adding a termination condition.
This overhaul removes the previous TTL system, and improves message passing by only requiring a total of V+E messages to be passed, where V is the number of drones and E is the number of connections.

The algorithm pretty much works by building a spanning tree within the network, and removes the path tracing and source routing headers, opting instead for simpler HashSets which can be merged (the Add trait is already implemented to do most of the job of merging replies), at the cost of requiring the drones to store some more information.

An option for broadcasting the addition of a new drone is also present, as is a "get path to X" feature which can be useful, for ex. in case of a crash to quickly find a valid connection in less time than a complete network flood.

Some options in the flooding protocol are optional and can be left unimplemented (for example, the "to" field can simply be ignored without causing any significant difference in returned replies, and drones don't have to broadcast themselves at init). There are also some performance considerations which can be chosen to be applied or not by groups.

+++ crates/wg_network/src/routing.rs
@@ -1,4 +1,4 @@
-pub type NodeId = u8;
+pub type NodeId = u64;
 
 #[derive(Debug, Clone)]
 pub struct SourceRoutingHeader {
